Try using boolean values for one-hot encoded features.

+++ autom8/categories.py
@@ -113,6 +113,5 @@ class OneHotEncoder:
         cols = []
         for i, series in enumerate(self.mapping):
             for value in series:
-                # SHOULD: Consider leaving the columns as booleans.
-                cols.append((X[:, i] == value).astype(int))
+                cols.append(X[:, i] == value)
         return np.stack(cols, axis=1)
